+++ src/demo/kotlin/com/loadingbyte/cinecred/demo/CommonProject.kt
@@ -73,7 +73,12 @@ private fun loadTemplateProject(modifyCsv: (Path) -> Unit = {}): Project =
     }
 
 
-val LOGO_PIC by lazy { useResourcePath("/logo.svg") { Picture.Loader.recognize(it)!!.apply { picture } } }
+val LOGO_PIC by lazy {
+    val tmpFile = Path(System.getProperty("java.io.tmpdir")).resolve("cinecred-logo.svg")
+    tmpFile.toFile().deleteOnExit()
+    tmpFile.writeLines(useResourceStream("/logo.svg") { it.bufferedReader().readLines() }.filter { "#000000" !in it })
+    Picture.Loader.recognize(tmpFile)!!.apply { picture }
+}
 
 val RAINBOW_TAPE: Tape by lazy {
     val tmpDir = Path(System.getProperty("java.io.tmpdir")).resolve("cinecred-rainbow")