+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Cinecred Changelog
   being baseline-aligned.
 - Added support for variable fonts.
 - Embedded pictures and videos may now be cropped, flipped, and rotated.
+- Embedded raster pictures may now have their blank space automatically cropped.
 - Embedded videos can now loop.
 - The preview resolution of embedded videos is now configurable.
 - The credits runtime can now be copied to the clipboard by clicking on it.